software/pi-mask-detection-master/requirements.txt
image==1.5.32
numpy==1.19.1
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
protobuf==3.13.0
tflite-runtime==2.1.0.post1
picamera==1.13
